requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Pillow==11.3.0
requests==2.32.5
beautifulsoup4==4.13.5
gunicorn==21.2.0